Day 90 of the Challenge
I think I'm okay in the styles of every section of my page that is why I started to add the media queries to make my portfolio webpage responsive. And right now I used em as media queries alike on the example, and I just recall how to convert em/rem to px and vice versa.